requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.30.0
authx==1.4.3
bcrypt==4.3.0
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
cryptography==45.0.4
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.13
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
filelock==3.18.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-oauth==0.16.1
hvac==2.3.0
identify==2.6.12
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
makefun==1.16.0
Mako==1.3.10
MarkupSafe==3.0.2
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pwdlib==0.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
python-status==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==5.3.1
requests==2.32.4
rsa==4.9.1
ruff==0.12.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
virtualenv==20.31.2
